refactor(parser): split large parser modules into smaller focused modules
Extract type, pattern, command, and import parsing into separate modules
to improve code organization and maintainability.

New modules:
- Type.hs (397 lines) - type parsing
- Pattern.hs (449 lines) - pattern parsing
- Cmd.hs (213 lines) - arrow command parsing
- Import.hs (204 lines) - module header and import parsing

Shared helpers (startsWithContextType, startsWithTypeSig, startsWithAsPattern,
isConLikeName, infixOperatorParserExcept, etc.) moved to Common.hs.

Module size changes:
- Expr.hs: 2307 → 1085 lines
- Decl.hs: 1688 → 1506 lines
- Common.hs: 822 → 950 lines

Progress counts unchanged. All 1257 tests pass.
