+++ components/aihc-parser/src/Aihc/Parser/Internal/FromTokens.hs
@@ -22,9 +22,12 @@ module Aihc.Parser.Internal.FromTokens
 where
 
 import Aihc.Parser.Internal.Common (TokParser, eofTok)
-import Aihc.Parser.Internal.Decl (declParser, importDeclParser, moduleHeaderParser)
-import Aihc.Parser.Internal.Expr (exprParser, patternParser, typeParser)
+import Aihc.Parser.Internal.Decl (declParser)
+import Aihc.Parser.Internal.Expr (exprParser)
+import Aihc.Parser.Internal.Import (importDeclParser, moduleHeaderParser)
 import Aihc.Parser.Internal.Module (moduleParser)
+import Aihc.Parser.Internal.Pattern (patternParser)
+import Aihc.Parser.Internal.Type (typeParser)
 import Aihc.Parser.Lex (LexToken)
 import Aihc.Parser.Syntax (Decl, Expr, ImportDecl, Module, ModuleHead, Pattern, Type)
 import Aihc.Parser.Types
